% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq2hdf5.R
\name{acq2hdf5}
\alias{acq2hdf5}
\title{Calls the acq2hdf5 command (python bioread package) to convert an acq file to an hdf5 file}
\usage{
acq2hdf5(
  acq_files,
  acq2hdf5_opts = "--compress=gzip",
  hdf5_output_dir = NULL,
  acq2hdf5_location = NULL
)
}
\arguments{
\item{acq_files}{Character vector of acq files to convert to hdf5.}

\item{acq2hdf5_opts}{Command line options passed to acq2hdf5. Default: "--missing-as=."}

\item{hdf5_output_dir}{Directory for converted hdf5 files. Will be created if it does not exist.}

\item{acq2hdf5_location}{The full path to the acq2hdf5 command. If not provided, this function will use the which/where command to look in the system path.}

\item{gzip_hdf5}{A logical (TRUE/FALSE) indicating whether to gzip hdf5 files after conversion. Default: TRUE}
}
\description{
This is an R wrapper around the acq2hdf5 python command. It also supports output directory specification
}
\details{
Note: the acq2hdf5 function is far faster than acq2hdf5 and also produces a much smaller file!
  In general, I would recommend that command for handling Acqknowledge files
}
