% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq2txt.R
\name{acq2txt}
\alias{acq2txt}
\title{Calls the acq2txt command (python bioread package) to convert an acq file to txt}
\usage{
acq2txt(
  acq_files,
  acq2txt_opts = "--missing-as=.",
  txt_output_dir = NULL,
  gzip_txt = TRUE,
  acq2txt_location = NULL
)
}
\arguments{
\item{acq_files}{Character vector of acq files to convert to txt.}

\item{acq2txt_opts}{Command line options passed to acq2txt. Default: "--missing-as=."}

\item{txt_output_dir}{Directory for converted txt files. Will be created if it does not exist.}

\item{gzip_txt}{A logical (TRUE/FALSE) indicating whether to gzip txt files after conversion. Default: TRUE}

\item{acq2txt_location}{The full path to the acq2txt command. If not provided, this function will use the which/where command to look in the system path.}
}
\description{
This is an R wrapper around the acq2txt python command. It also supports output directory specification and gzip'ing files
}
\details{
Note: the acq2hdf5 function is far faster than acq2txt and also produces a much smaller file!
  In general, I would recommend that command for handling Acqknowledge files
}
