% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{biopac_hdf5_to_dataframe}
\alias{biopac_hdf5_to_dataframe}
\title{Read a BIOPAC hdf5 file as a data.frame}
\usage{
biopac_hdf5_to_dataframe(
  hdf5file,
  upsample_to_max = TRUE,
  ttl_to_dec = TRUE,
  ttl_columns = NULL
)
}
\arguments{
\item{hdf5file}{HDF5 file containing converted BIOPAC data (using acq2hdf5)}

\item{upsample_to_max}{logical (TRUE/FALSE) indicating whether to upsample slower channels to the sampling
rate of the fastest channel. If \code{TRUE}, upsampling is performed on slower channels through linear interpolation.
If \code{FALSE}, NA values are inserted for unsampled time points.}

\item{ttl_to_dec}{logical indicating whether to convert binary digital input channels to decimal TTL (parallel port)
codes. These codes are used to synchronize timing between behavioral and physiological data streams.}

\item{ttl_columns}{A numeric vector or character vector indicating which columns within the imported dataset
contain the digital inputs that should be converted to decimal TTL codes. If not specified, channels
starting with Digital.input will be used. There should be exactly 8 channels for a 1-byte (0-255) TTL code.}
}
\description{
After converting an acq file using acq2hdf5, this function converts the hdf5 file to a data.frame
}
\details{
This function uses the data.table package to build the data.frame of physiological data because this package
  is particularly good at working with large datasets without a lot of copy-in-memory steps that can slow down
  processing and explode the peak RAM demand.

  The function also supports upsampling of all data onto the same time resolution as the fastest
  sampling rate for any channel. This is governed by the \code{upsample_to_max} argument
}
