% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\docType{class}
\name{Block-class}
\alias{Block-class}
\title{Class Block}
\arguments{
\item{input.blocks}{{character} Previously attached blocks}

\item{output.blocks}{{character} Block to move to end of event}
}
\description{
Class Block
}
