% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{block.validity}
\alias{block.validity}
\title{Checks validity of block object}
\usage{
block.validity(object)
}
\arguments{
\item{object}{Block object}
}
\value{
true boolean or list of errors
}
\description{
Checks that the input block is a block object or null.
Checks that the output block is a block object or null.
}
