% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_functions.R
\name{compute_ttl_onsets}
\alias{compute_ttl_onsets}
\title{Recode TTL code vector to reflect changes in the value}
\usage{
compute_ttl_onsets(ttl_vec, lazy_ttl = 2, zero_code = 0)
}
\arguments{
\item{ttl_vec}{a vector of integer TTL codes reconstructed from the acq file}

\item{lazy_ttl}{a non-negative integer indicating the lowest number of samples that could
reflect a genuine change in a TTL code. Default: 2}

\item{zero_code}{what code reflects that all pins are off (no active code). If you have a stuck
parallel port pin, set \code{zero_code} to that value to prevent flagging onsets of that value
as events.}
}
\description{
This functions helps to identify the key time markers when TTL codes onset,
  which is hard to detect using the raw vector alone (since the code will have some duration)
}
\details{
The parallel port cannot truly change all 8 pins simultaneously for a new code.
  If the speed of the pins changing falls on a sampling boundary on the BIOPAC hardware, this can lead
  to false codes for very brief periods of time (~2ms). Thus, identify rapid code changes, then
  interrogate around these, replacing the dubious code with the code that 'sticks' subsequently.
  This decision means that if any pin has changed, then the experiment has executed a new TTL code.

  For example, at 1000Hz, @\code{lazy_ttl} = 2 would flag any codes with duration of 2ms or less as
  suspicious and would replace these with the longer, genuine code that immediately follows.
}
\author{
Michael Hallquist
}
