% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{connect.blocks}
\alias{connect.blocks}
\title{Connects blocks}
\usage{
connect.blocks(input.block, output.block)
}
\arguments{
\item{input.block}{Input block}

\item{output.block}{Output block}
}
\value{
list
}
\description{
Connects two blocks together as a directed graph style
}
