% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_peaks.R
\name{correct_peaks}
\alias{correct_peaks}
\title{Adjust a set of detected peaks to coincide with local signal maxima.}
\usage{
correct_peaks(
  sig,
  peak_inds,
  search_radius,
  smooth_window_size,
  peak_dir = "compare"
)
}
\arguments{
\item{sig}{The 1d signal vector}

\item{peak_inds}{Vector of the original peak indices}

\item{search_radius}{The radius within which the original peaks may be shifted.}

\item{smooth_window_size}{The window size of the moving average filter applied on the
signal. Peak distance is calculated on the difference between
the original and smoothed signal.}

\item{peak_dir}{optional string denoting the expected peak direction:
'up' or 'down', 'both', or 'compare'.}
}
\value{
shifted_peak_inds vector of the corrected peak indices.
}
\description{
Adjust a set of detected peaks to coincide with local signal maxima.
}
\details{
Functionality of \code{peak_dir} param:
       - If 'up', the peaks will be shifted to local maxima.
       - If 'down', the peaks will be shifted to local minima.
       - If 'both', the peaks will be shifted to local maxima of the
         rectified signal.
       - If 'compare', the function will try both 'up' and 'down'
         options, and choose the direction that gives the largest mean
         distance from the smoothed signal.
}
