% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{create.block}
\alias{create.block}
\title{Create a block object}
\usage{
create.block(block.name, callback = function(l) {
    return(l)
})
}
\arguments{
\item{block.name}{Name of block}

\item{callback}{callback function (Feed-forward pass-through if none)}
}
\value{
block object
}
\description{
Block objects allow for easy maniputlation of dataflow.
Each block object is given an input, output, and function.
The input represents where the block's inputs come from (NULL if none).
The output represents where the block's outputs go to (NULL if none).
The callback function is the actual manipulation of data.
}
