% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-logger.R
\name{create.log}
\alias{create.log}
\title{Creates a log object that outputs to the console.}
\usage{
create.log(name, level = 20)
}
\arguments{
\item{name}{Name of the logger.}

\item{level}{Level for logger to operate at.}
}
\description{
Creates a log object that outputs to the console.
}
