% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_digital_timeseries_r.R
\name{downsample_digital_timeseries_r}
\alias{downsample_digital_timeseries_r}
\title{here's a pure R version of downsampling a categorical/digital timeseries}
\usage{
downsample_digital_timeseries_r(x, downsamp, demote_zeros = TRUE)
}
\arguments{
\item{x}{input digital (categorical) timeseries vector}

\item{downsamp}{downsampling factor (integer > 1)}

\item{demote_zeros}{If TRUE, don't allow zero to be the retained value within a downsampled chunk}
}
\description{
here's a pure R version of downsampling a categorical/digital timeseries
}
