% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_functions.R
\name{downsample_physio}
\alias{downsample_physio}
\title{downsample_physio}
\usage{
downsample_physio(
  ep.physio,
  downsample_factor = 1,
  digital_channels = c("ttl_code", "ttl_onset", "Digital.*"),
  method = "decimate"
)
}
\arguments{
\item{ep.physio}{An ep.physio object created by read_acq}

\item{downsample_factor}{An integer factor used to subsample data}

\item{digital_channels}{Column names or positions containing digital channels, These will be downsampled using the
\code{downsample_digital_timeseries} function to use the within-chunk mode, rather than blind subsampling.}

\item{method}{How to downsample the signal. The default is \code{"decimate"}, which calls \code{signal::decimate}.
This applies a low-pass filter before downsampling to avoid aliasing. The alternative is \code{"subsample"}, which
simply takes every nth sample from the original time series.}
}
\description{
This function reduces the
}
