% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg_qa.R
\name{ecg_stripplot}
\alias{ecg_stripplot}
\title{A convenience function for plotting ECG time series and QRS detection}
\usage{
ecg_stripplot(
  ecg_df,
  beat_df = NULL,
  freq = 1000,
  max_rows = 5,
  spike_annotations = c("gqrs"),
  ibi_splines = c("gqrs"),
  spi = 2,
  width = 12,
  height = 10,
  out_pdf = "ecg_stripplot.pdf"
)
}
\arguments{
\item{ecg_df}{a data.frame containing ECG data. Must have at least the columns 'time' and 'ecg',
which are used for x and y axes, respectively.}

\item{beat_df}{an optional data.frame containing beat events to plot against the trace}

\item{freq}{the frequency of the data in Hz}

\item{max_rows}{the number of rows on each page of the PDF}

\item{spike_annotations}{which QRS algorithm columns to plot as points on the stripplot from beat_df}

\item{ibi_splines}{which QRS algorithm RR columns to plot as splines against the ECG data}

\item{spi}{How many seconds of data per inch to plot on the page. Default is 2, which allows one to
resolve each beat by eye. Higher values (e.g., 10) may be useful for a summary view and to look for artifacts.}

\item{width}{the width of the PDF output in inches. Default: 12}

\item{height}{the height of the PDF output in inches. Default: 10}

\item{out_pdf}{the filename (and optionally, path) of the PDF to be created}
}
\description{
A convenience function for plotting ECG time series and QRS detection
}
