% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_add_aois}
\alias{ep.eye_add_aois}
\title{Add AOI information to gaze data}
\usage{
ep.eye_add_aois(
  ep.eye,
  indicator,
  extract_coords,
  extract_labs,
  split_coords,
  tag_raw = FALSE
)
}
\arguments{
\item{ep.eye}{an ep.eye object}

\item{indicator}{a regex containing string to search for across messages for AOI information}

\item{extract_coords}{a regex for extracting AOI coordinates}

\item{extract_labs}{a regex for extracting AOI labels}

\item{split_coords}{string denoting what separator is being used to split AOI coordinates}

\item{tag_raw}{logical. tag raw data with AOI specific information? Default to FALSE.}
}
\description{
Add AOI information to gaze data
}
