% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_setup_proc_config_helpers.R
\name{ep.eye_build_msg_seq}
\alias{ep.eye_build_msg_seq}
\title{Build out expected message sequences within config file.}
\usage{
ep.eye_build_msg_seq(config, dt = NULL)
}
\arguments{
\item{config}{Named list of configuration options read in by \code{validate_exp_yaml}}

\item{dt}{Descriptive text to print after running. Defaults to NULL (silent).}
}
\value{
Nested list with configuration options.
}
\description{
When a message sequence check is requested, the user specifies event-general start and end message sequences, with the message sent during the event being unique to the block and event. This function attempts to combine the general and specific into the msg_seq field of msg_parse options, which gives block/event-level specificity on the exact expected sequence of messages to check.
}
\author{
Nate Hall
}
