% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events_helpers.R
\name{ep.eye_check_msg_order}
\alias{ep.eye_check_msg_order}
\title{Check ordering of messages}
\usage{
ep.eye_check_msg_order(ep.eye, msg_seq, msg_check, eventnum)
}
\arguments{
\item{ep.eye}{ep.eye object that has been previously initialized and tidied}

\item{msg_seq}{List of optional message sequence arguments passed in config file. Can contain \code{msg_start}, \code{msg_end}, \code{eval_middle}, \code{ordered}. See the ep.eye_config vignette for details on these fields.}

\item{msg_check}{data.frame generated from \code{ep.eye_check_requested_msg}}

\item{eventnum}{numeric value of the event number to check}
}
\value{
df msg_check with seq_match appended as column name.
}
\description{
Check ordering of messages
}
\author{
Nate Hall
}
