% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events_helpers.R
\name{ep.eye_check_requested_msg}
\alias{ep.eye_check_requested_msg}
\title{Compare requested (from config) and extracted (from ep.eye$raw) messages for completeness within a single event}
\usage{
ep.eye_check_requested_msg(ep.eye, eventnum)
}
\arguments{
\item{ep.eye}{ep.eye object that has been previously initialized and tidied}

\item{eventnum}{numeric value of the event number to check}
}
\value{
df a data.frame with columns c("requested", "match", and "extracted") which provide a mapping between all messages that should be within a given event and the actual message that is extracted from the .edf file. all(match) == TRUE indicates that all requested messages are contained within the et.msgs in the .edf file (good thing).
}
\description{
Compare requested (from config) and extracted (from ep.eye$raw) messages for completeness within a single event
}
\author{
Nate Hall
}
