% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_cleanup.R
\name{ep.eye_cleanup}
\alias{ep.eye_cleanup}
\title{Eye Data Preprocessing and Cleanup}
\usage{
ep.eye_cleanup(ep.eye, config, globals, header = NULL)
}
\arguments{
\item{ep.eye}{An object containing the eye tracking data to be cleaned up.}

\item{config}{config file}

\item{globals}{A list of global parameters defined in the config file.}

\item{header}{A string used to set the chunk header for the logging. Default is NULL.}
}
\value{
Returns the cleaned up eye tracking data with the class "ep.eye.preproc".
}
\description{
This function processes eye tracking data for use in event-related potential (ERP) studies.
It generates event-locked timing, removes raw data to minimize object size, tags the data
with a class to indicate successful preprocessing, saves the preprocessed data, and manages log files.
}
\examples{
\dontrun{
cleaned_data <- ep.eye_cleanup(raw_eye_data, config$definitions$eye$global, "6. Cleanup and export ep.eye")
}
}
\seealso{
\code{\link{ep.eye_tag_event_time}}, \code{\link{ep.eye_save_preproc}}
}
