% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_setup_proc_config_helpers.R
\name{ep.eye_default_options}
\alias{ep.eye_default_options}
\title{Get Default Eye-Tracking Processing Options}
\usage{
ep.eye_default_options(edf_raw)
}
\value{
A named list with default options for the following fields:
\itemize{
  \item \strong{global:} General processing options, including:
  \itemize{
    \item \code{base_dir} (character): The base directory, defaults to the current working directory.
    \item \code{prefix} (character): A regular expression pattern for the file prefix, defaults to "\\d{2,3}".
    \item \code{save_steps} (logical): Flag to save intermediate processing steps, defaults to \code{TRUE}.
    \item \code{preproc_out} (character): Directory for preprocessed output, defaults to "preproc_out" in the current working directory.
    \item \code{log} (logical): Flag to enable logging, defaults to \code{FALSE}.
    \item \code{remove_raw} (logical): Flag to return raw data, defaults to \code{FALSE}.
  }
  \item \strong{initialize:} Initialization options, including:
  \itemize{
    \item \code{expected_edf_fields} (character vector): Expected EDF file fields, defaults to \code{c("raw", "sacc", "fix", "blinks", "msg", "input", "button", "info", "asc_file")}.
    \item \code{unify_gaze_events} (list): Gaze event unification options, defaults to:
    \itemize{
      \item \code{gaze_events} (character vector): Gaze events to unify, defaults to \code{c("sacc", "fix", "blink")}.
      \item \code{confirm_correspondence} (logical): Flag to confirm correspondence of gaze events, defaults to \code{FALSE}.
    }
  }
  \item \strong{msg_parse:} Message parsing options, defaults to \code{NULL}.
  \item \strong{gaze_preproc:} Gaze preprocessing options, including:
  \itemize{
    \item \code{aoi} (list): Area of interest (AOI) options, defaults to:
    \itemize{
      \item \code{indicator} (character): Indicator for AOI, defaults to "!V IAREA RECTANGLE".
      \item \code{extraction_method} (character): Method for AOI extraction, defaults to "regex".
      \item \code{extraction_coords} (character): Regular expression for extracting coordinates, defaults to "\\d{3,4} \\d{3,4} \\d{3,4} \\d{3,4}".
      \item \code{extract_labs} (character): Regular expression for extracting labels, defaults to "[a-z]+$".
      \item \code{split_coords} (character): Character used to split coordinates, defaults to a single space.
      \item \code{tag_raw} (logical): Flag to tag raw data, defaults to \code{FALSE}.
    }
    \item \code{downsample} (list): Downsampling options, defaults to:
    \itemize{
      \item \code{downsampled_freq} (numeric): Frequency to downsample to, defaults to \code{50} Hz.
      \item \code{method} (character): Method for downsampling, defaults to "mean".
    }
  }
  \item \strong{pupil_preproc:} Pupil preprocessing options, including:
  \itemize{
    \item \code{blink_corr} (list): Blink correction options, defaults to:
    \itemize{
      \item \code{ms_before} (numeric): Time before blink in milliseconds, defaults to \code{150}.
      \item \code{ms_after} (numeric): Time after blink in milliseconds, defaults to \code{150}.
    }
    \item \code{filter} (list): Filtering options, defaults to:
    \itemize{
      \item \code{method} (character): Filtering method, defaults to "movingavg".
      \item \code{window_length} (numeric): Length of the filtering window, defaults to \code{20}.
    }
    \item \code{interpolate} (list): Interpolation options, defaults to:
    \itemize{
      \item \code{algor} (character): Interpolation algorithm, defaults to "linear".
      \item \code{maxgap} (numeric): Maximum gap for interpolation in milliseconds, defaults to \code{1000}.
    }
    \item \code{baseline_correction} (list): Baseline correction options, defaults to:
    \itemize{
      \item \code{method} (character): Method for baseline correction, defaults to "subtract".
      \item \code{dur_ms} (numeric): Duration in milliseconds for baseline correction, defaults to \code{100}.
      \item \code{center_on} (character): Event to center baseline correction on, defaults to "DISPLAY_ON".
    }
    \item \code{downsample} (list): Downsampling options, defaults to:
    \itemize{
      \item \code{downsampled_freq} (numeric): Frequency to downsample to, defaults to \code{20} Hz.
      \item \code{method} (character): Method for downsampling, defaults to "mean".
    }
  }
}
}
\description{
Returns a hierarchical list containing the default options for the eye-tracking data processing pipeline.
}
\examples{
defaults <- ep.eye_default_options()
str(defaults)

}
\author{
Nate Hall
}
