% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_display_steps.R
\name{ep.eye_display_steps}
\alias{ep.eye_display_steps}
\title{Display Steps for Eye-Tracking Data Processing}
\usage{
ep.eye_display_steps()
}
\value{
None. This function prints the steps to the console.
}
\description{
This function displays the various steps involved in processing eye-tracking data.
Each step includes a keyword, the function invoked, example arguments for the function, and a description of the step.
}
\section{Steps}{

\describe{
  \item{1. Keyword: config}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_setup_proc_config}}
      \item{Function Arguments:}{\code{"edf_raw, file_path, header"}}
      \item{Description:}{Processes user-supplied configuration yaml}
    }
  }
  \item{2. Keyword: init}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_initialize}}
      \item{Function Arguments:}{\code{"config_path"}}
      \item{Description:}{Initializes the eye-tracking data using configurations from a YAML file.}
    }
  }
  \item{3. Keyword: parse}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_setup_proc_config}}
      \item{Function Arguments:}{\code{"event_functions_path"}}
      \item{Description:}{Parses task-specific events from the eye-tracking data.}
    }
  }
  \item{4. Keyword: gaze_preproc}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_setup_proc_config}}
      \item{Function Arguments:}{\code{"aoi, downsample"}}
      \item{Description:}{Preprocesses gaze data including filtering, downsampling, etc.}
    }
  }
  \item{5. Keyword: pupil_preproc}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_setup_proc_config}}
      \item{Function Arguments:}{\code{"blink_corr, filter, interpolate, baseline_correction"}}
      \item{Description:}{Preprocesses pupil data including blink correction, filtering, etc.}
    }
  }
  \item{6. Keyword: cleanup}{
    \describe{
      \item{Function Invoked:}{\code{ep.eye_setup_proc_config}}
      \item{Function Arguments:}{\code{"globals"}}
      \item{Description:}{Cleans up processed data and prepares it for export.}
    }
  }
}
}

\examples{
ep.eye_display_steps()

}
