% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_downsample}
\alias{ep.eye_downsample}
\title{Downsample gaze and/or pupil data}
\usage{
ep.eye_downsample(
  df,
  sample.rate = 1000,
  downsampled_freq = 20,
  digital_channels = c("eventn", "saccn", "fixn", "blinkn", "block_trial"),
  analog_channels = c("xp", "yp", "ps"),
  char_channels = c("et.msg", "block", "event"),
  method = "mean"
)
}
\arguments{
\item{df}{data.table to downsample (usually this is ep.eye$raw)}

\item{sample.rate}{sampling rate of the eyetracker in Hz}

\item{downsampled_freq}{Frequency of the downsampled data indicating degree of downsampling. It is important to be mindful of your sampling rate when selecting this value. Defaults to 20 (which moves a second of recording at 1000Hz to 20 measurements).}

\item{digital_channels}{Character vector of columns in df that are integer values that should not be combined but blocked by.}

\item{analog_channels}{Character vector of columns in df that are continually varying and should be summarized within the downsampling procdure (incl x and y gaze position and pupil size)}

\item{char_channels}{Character vector of columns in df that represent values that should be combined if there are unique values within a downsampling block. If there are multiples within a downsampling block will paste them together with " || "}

\item{method}{String "mean" or "subsample" determining whether to perform subsampling (keep every n measurement) or mean-based downsampling.}
}
\description{
Downsample gaze and/or pupil data
}
