% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_gen_aoi_ref}
\alias{ep.eye_gen_aoi_ref}
\title{Generate AOI-reference object}
\usage{
ep.eye_gen_aoi_ref(
  ep.eye,
  indicator,
  extract_coords,
  extract_labs,
  split_coords,
  dt
)
}
\arguments{
\item{ep.eye}{an ep.eye object}

\item{indicator}{a regex containing string to search for across messages for AOI information}

\item{extract_coords}{a regex for extracting AOI coordinates}

\item{extract_labs}{a regex for extracting AOI labels}

\item{split_coords}{string denoting what separator is being used to split AOI coordinates}
}
\description{
Generate AOI-reference object
}
