% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_get_session_length}
\alias{ep.eye_get_session_length}
\title{Tag ep.eye metadata with the length of the recording session}
\usage{
ep.eye_get_session_length(ep.eye)
}
\arguments{
\item{ep.eye}{Initialized ep.eye structure.}
}
\value{
ep.eye Modified ep.eye structure with metadata$recording time denoting session recording length in seconds.
}
\description{
Tag ep.eye metadata with the length of the recording session
}
\author{
Nate Hall
}
