% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_inherit_btw_ev}
\alias{ep.eye_inherit_btw_ev}
\title{split off function for conversion of between trial messages to within}
\usage{
ep.eye_inherit_btw_ev(ep.eye, inherit_btw_ev, dt)
}
\description{
split off function for conversion of between trial messages to within
}
