% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_interp_pupil}
\alias{ep.eye_interp_pupil}
\title{Interpolate over missing pupil data}
\usage{
ep.eye_interp_pupil(ep.eye, algor = "linear", maxgap = 1000)
}
\arguments{
\item{ep.eye}{ep.eye object containing ps_smooth as a column in pupil data.}

\item{algor}{preferred interpolation algorithm, see imputeTS::na_interpolation for options}

\item{maxgap}{Maximum amount of missing *time (in ms)* to interpolate over. Anything above this will be left NA. Converts internally from time to nmeasurements if data is not sampled at 1000Hz.}
}
\description{
Interpolate over missing pupil data
}
