% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events.R
\name{ep.eye_parse_events}
\alias{ep.eye_parse_events}
\title{Parse task config file, extracting essential task information}
\usage{
ep.eye_parse_events(
  ep.eye,
  extraction_method,
  extract_event_func_path,
  csv_path,
  msg_seq,
  header = NULL
)
}
\arguments{
\item{ep.eye}{ep.eye object that has been previously initialized and tidied}

\item{header}{String for section header. Defaults to "4. Parse task events:"}
}
\value{
ep.eye appended with important event-level information included.
}
\description{
Parse task config file, extracting essential task information
}
\author{
Nate Hall
}
