% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_gaze.R
\name{ep.eye_preprocess_gaze}
\alias{ep.eye_preprocess_gaze}
\title{Preprocess gaze data}
\usage{
ep.eye_preprocess_gaze(ep.eye, config, aoi, downsample, header = NULL)
}
\arguments{
\item{ep.eye}{Path to a single \code{.edf} file using \code{read_edf()}.}

\item{config}{config file}

\item{aoi}{list of AOI options}

\item{downsample}{list of downsampling options. Contains $factor, and $method.}

\item{task}{Character value with task name. Contains $indicator, $extract_coords, $extract_labs, $split_coords}
}
\value{
ep.eye an ep.eye object that has gaze data preprocessed.
}
\description{
Performs AOI tagging (within gaze event fields) and downsampling (within raw gaze data) of gaze data.
}
\author{
Nate Hall
}
