% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_pupil.R
\name{ep.eye_preprocess_pupil}
\alias{ep.eye_preprocess_pupil}
\title{Preprocess Pupil Data}
\usage{
ep.eye_preprocess_pupil(
  ep.eye,
  blink_corr,
  filter,
  interpolate,
  baseline_correction,
  downsample,
  header = NULL
)
}
\arguments{
\item{ep.eye}{A data object containing eye-tracking information.}

\item{blink_corr}{A list containing parameters for blink correction: `ms_before` and `ms_after` specifying the milliseconds before and after a blink to be corrected.}

\item{filter}{A list containing parameters for filtering: `method` and `window_length` specifying the method and window length for filtering.}

\item{interpolate}{A list containing parameters for interpolation: `algor` and `maxgap` specifying the algorithm and maximum gap for interpolation.}

\item{baseline_correction}{A list containing parameters for baseline correction: `method`, `dur_ms`, and `center_on` specifying the method, duration in milliseconds, and the centering method for baseline correction.}

\item{downsample}{A list containing parameters for downsampling: `downsampled_freq` and `method` specifying the downsampled frequency and method for downsampling.}

\item{header}{An optional header parameter. If NULL, no header is used.}
}
\value{
A data object containing the preprocessed pupil data.
}
\description{
This function preprocesses pupil data by extending blinks, filtering, interpolating, performing baseline correction, and downsampling the data.
}
\examples{
# Assuming `ep.eye`, `blink_corr`, `filter`, `interpolate`, `baseline_correction`, and `downsample` are defined:
preprocessed_data <- ep.eye_preprocess_pupil(ep.eye, blink_corr, filter, interpolate, baseline_correction, downsample)
}
