% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_process_subject.R
\name{ep.eye_process_subject}
\alias{ep.eye_process_subject}
\title{Read and process a single .edf file.}
\usage{
ep.eye_process_subject(edf_raw, config_path, step = NULL, ...)
}
\arguments{
\item{config_path}{Path to corresponding .yml configuration file with processing instructions. Instructions on how to effectively set up a configuration file can be found [HERE].}

\item{step}{complete preprocessing up to which step?}

\item{...}{Optional arguments to pass to \code{read_edf.R} function.}

\item{file}{Path to the .edf file to process.}
}
\value{
A fully processed ep.eye object. [DETAILS HERE]
}
\description{
This is the main worker function that processes a single subject through the ep.eye preprocessing pipeline. It takes as arguments the paths to an .edf file and corresponding configuration YAML file and exports a preprocessed ep.eye object for a single subject.
}
\details{
This function is ideally used within the \code{ep_batch_process_eye.R}

data.table notes: one annoying challenge is making sure that data.table has openMP support enabled. See https://github.com/Rdatatable/data.table/issues/5419#issuecomment-1465023581 and https://stackoverflow.com/questions/76374014/use-openmp-on-m2-mac-with-r-and-data-table. You will need to download a few packages and update Makevars and your bashrc file.
}
\examples{
 \dontrun{
   file <- "/proj/mnhallqlab/studies/NeuroMAP/s3_data/Neighborhood_PSU/Eye/004_AZ_Neighborhood_Eye.edf"
   config_path <- "/proj/mnhallqlab/studies/NeuroMAP/s3_data_ep_specs/yaml/Neighborhood_PSU.yaml"
   ep.eye <- read_process_eye(file, config_path)
 }
}
\author{
Nate Hall
}
