% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_raw_sample_continuity_check}
\alias{ep.eye_raw_sample_continuity_check}
\title{Tag ep.eye metadata with the length of the recording session}
\usage{
ep.eye_raw_sample_continuity_check(ep.eye)
}
\arguments{
\item{ep.eye}{Initialized ep.eye structure.}
}
\value{
ep.eye Modified ep.eye structure with metadata$missing_measurements, either NULL if all timepoints are strictly accounted for or a data.table documenting "runs" of missing data.
}
\description{
Tag ep.eye metadata with the length of the recording session
}
\author{
Nate Hall
}
