% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_rm_crinfo}
\alias{ep.eye_rm_crinfo}
\title{Remove useless cr.info column.}
\usage{
ep.eye_rm_crinfo(ep.eye)
}
\arguments{
\item{ep.eye}{An ep.eye object.}
}
\value{
ep.eye
}
\description{
Removes cr.info column if it contains unhelpful information
}
\details{
From Eyelink 100 manual: 4.9.3.1 Samples Recorded in Corneal Reflection Mode: If the data file being processed was recorded using corneal reflection mode, each sample line has an added 3 (monocular) or 5 (binocular) character fields after all other fields (including resolution and velocity if enabled). These fields represent warning messages for that sample relating to the corneal reflection processing.
         MONOCULAR Corneal Reflection (CR) Samples
           - "..." if no warning for sample
           - first character is "I" if sample was interpolated second character is "C" if CR missing
           - third character is "R" if CR recovery in progress
}
