% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_rm_impossible}
\alias{ep.eye_rm_impossible}
\title{Remove values outside of screen dimensions}
\usage{
ep.eye_rm_impossible(ep.eye, dt = NULL)
}
\arguments{
\item{ep.eye}{an ep.eye object}

\item{dt}{descriptive text to print to log file, defaults to NULL.}
}
\description{
Remove values outside of screen dimensions
}
