% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_setup_proc_config_helpers.R
\name{ep.eye_set_config_definitions}
\alias{ep.eye_set_config_definitions}
\title{Implement Default Definitions}
\usage{
ep.eye_set_config_definitions(config, field)
}
\arguments{
\item{config}{}
}
\description{
Implement Default Definitions
}
