% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_setup_proc_config.R
\name{ep.eye_setup_proc_config}
\alias{ep.eye_setup_proc_config}
\title{Read and setup processing configuration}
\usage{
ep.eye_setup_proc_config(edf_raw, config_path, header = NULL)
}
\arguments{
\item{config_path}{Path to corresponding .yml configuration file with processing instructions. Instructions on how to effectively set up a configuration file can be found [HERE].}

\item{file}{Path to the .edf file to process.}
}
\value{
Nested list with processing options pulled from .yml configuration file.
}
\description{
Validates and imports processing configuration options into the environment for later use. Implements default configuration options if missing from YAML.
}
\details{
Also generates new log file using \code{sink()}, and sets  up a new directory for preprocessed output if it does not already exist.
}
\examples{
 \dontrun{
   config <- setup_proc_configs("/proj/mnhallqlab/studies/NeuroMAP/s3_data/Neighborhood_PSU/eye/002_HS_Neighborhood_Eye.edf", "/proj/mnhallqlab/studies/NeuroMAP/s3_data_ep_specs/yaml/Neighborhood_PSU.yaml")
 }
}
\author{
Nate Hall

importFrom stringr str_extract
importFrom yaml read_yaml
importFrom nate.utils print_tree
}
