% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_setup_structure}
\alias{ep.eye_setup_structure}
\title{Setup ep.eye structure}
\usage{
ep.eye_setup_structure(eye, task = NULL)
}
\arguments{
\item{eye}{Loaded .edf file from \code{read_edf}}

\item{task}{Task name (character)}
}
\value{
ep.eye Initialized ep.eye structure. [DETAILS HERE].
}
\description{
Setup ep.eye structure
}
\author{
Nate Hall
}
