% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_shift_timing}
\alias{ep.eye_shift_timing}
\title{Shift timing of ep.eye object to 0 start point}
\usage{
ep.eye_shift_timing(ep.eye, dt = NULL)
}
\arguments{
\item{ep.eye}{An ep.eye object.}

\item{dt}{descriptive text to print to log file, defaults to NULL.}
}
\value{
ep.eye ep.eye structure with all relevant elements recoded to 0 start point. Original start point is stored in ep.eye$metadata$t_start
}
\description{
Shift timing of ep.eye object to 0 start point
}
\author{
Nate Hall
}
