% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_unify_gaze_events}
\alias{ep.eye_unify_gaze_events}
\title{Unify gaze events with raw data}
\usage{
ep.eye_unify_gaze_events(ep.eye, gaze_events = c("sacc", "fix", "blink"))
}
\arguments{
\item{ep.eye}{An initialized ep.eye object}

\item{gaze_events}{Character vector of gaze_events to unify with raw. Defaults to unifying sacc, fix, and blink but can be set to any subset of these.}
}
\value{
ep.eye ep.eye structure that has been tagged with gaze event numbers and validated for correspondence between raw and gaze event fields.
}
\description{
When loaded into the environment, an .edf file will be parsed with fields corresponding to $raw, $sacc, $fix, and $blink. Saccades, fixations, and blinks are called "gaze events" in the ep.eye nomenclature and denote the presence of an event of significance within gaze data. This function unifies raw data with gaze events by generating unique gaze event numbers, merges them to raw data and validates that timestamps from raw and gaze event fields are in correspondence.
}
\author{
Nate Hall
}
