% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_unify_raw_msg}
\alias{ep.eye_unify_raw_msg}
\title{Add et.msg column to raw data}
\usage{
ep.eye_unify_raw_msg(ep.eye)
}
\arguments{
\item{ep.eye}{An ep.eye object.}
}
\value{
ep.eye
}
\description{
Adds within-event messages from \code{ep.eye$msg} to \code{ep.eye$raw}.
}
