% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events_helpers.R
\name{ep.eye_validate_msg_seq}
\alias{ep.eye_validate_msg_seq}
\title{split off function for checking message sequence.}
\usage{
ep.eye_validate_msg_seq(ep.eye, msg_seq, dt)
}
\description{
split off function for checking message sequence.
}
