% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_setup_proc_config_helpers.R
\name{ep.list.tree}
\alias{ep.list.tree}
\title{Print list structure}
\usage{
ep.list.tree(
  struct,
  depth = -1,
  numbers = FALSE,
  maxlen = 10000,
  maxcomp = 30,
  attr.print = TRUE,
  front = "",
  fill = "-",
  name.of,
  size = FALSE,
  digits = 5
)
}
\description{
this is a very slight ammendment to the Hmisc package for easy display of lists: http://math.furman.edu/~dcs/courses/math47/R/library/Hmisc/html/list.tree.html
}
