% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{find.unprocessed}
\alias{find.unprocessed}
\title{Finds Unprocessed files based on file names}
\usage{
find.unprocessed(dir.raw, dir.processed, input.file.extension = "")
}
\arguments{
\item{dir.raw}{Directory location for the raw directory.}

\item{dir.processed}{Directory location for the processed directory.}

\item{input.file.extension}{File extension to look for (empty string to disregard (standard))}
}
\value{
raw file names
}
\description{
Finds files that exist in a raw directory, which do not have a counterpart
in a processed directory.  Files names are expected to have the same file name
i.e. foo_bar.fb <-> foo_bar.bf
}
