% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{get.block.name}
\alias{get.block.name}
\title{Get name of block object}
\usage{
get.block.name(block)
}
\arguments{
\item{block}{Block object}
}
\value{
name of block
}
\description{
Get name of block object
}
