% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{get.callback}
\alias{get.callback}
\title{Get callback function of block object}
\usage{
get.callback(block)
}
\arguments{
\item{block}{Block object}
}
\value{
callback function of block
}
\description{
Get callback function of block object
}
