% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{get.input.blocks}
\alias{get.input.blocks}
\title{Get input of block object}
\usage{
get.input.blocks(block)
}
\arguments{
\item{block}{Block object}
}
\value{
input of block
}
\description{
Get input of block object
}
