% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{get.output.blocks}
\alias{get.output.blocks}
\title{Get output of block object}
\usage{
get.output.blocks(block)
}
\arguments{
\item{block}{Block object}
}
\value{
output of block
}
\description{
Get output of block object
}
