% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-config.R
\name{get.system}
\alias{get.system}
\title{Get System Info}
\usage{
get.system()
}
\description{
Retrieves system info defined by R.
Info includes:
- Operating System (OSX or Windows)
}
\details{
All information is saved to global param on NeuroMap environment: SYSTEM
}
