% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import.edf}
\alias{import.edf}
\title{Imports EDF files}
\usage{
import.edf(dir)
}
\arguments{
\item{dir}{base directory for raw and proc directories.}
}
\value{
list of edf file locations and edf data.
}
\description{
Imports EDF (eyetracker) data and creates asc as well as read the data
into dataframes.
}
