% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import.mat}
\alias{import.mat}
\title{Imports MAT files}
\usage{
import.mat(dir)
}
\arguments{
\item{dir}{base directory for raw and proc directories.}
}
\value{
list of edf file locations and edf data.
}
\description{
Imports MAT (matlab) data and creates asc as well as read the data
into dataframes.
}
