% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import.shrooms}
\alias{import.shrooms}
\title{Imports shrooms data}
\usage{
import.shrooms(dir)
}
\arguments{
\item{dir}{directory to import from}
}
\value{
list of asc and mat data
}
\description{
Imports shrooms data as dataframes.
}
