% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-config.R
\name{import.yaml}
\alias{import.yaml}
\title{Read YAML Config File}
\usage{
import.yaml(file, configuration = "default")
}
\arguments{
\item{file}{YAML file path}

\item{configuration}{(default = "default") Configuration of YAML to read}
}
\value{
list of YAML configurations
}
\description{
Read YAML configuration file defined.
All YAML document objects are returned as a list.
Configurations can be used to defined different configurations
from the YAML file.
All output is saved to global parameter in NeuroMap environment: CONFIG
}
