% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_acq}
\alias{import_acq}
\title{Import Acqknowledge ACQ files into data.frames}
\usage{
import_acq(
  acq_files,
  hdf5_output_dir = NULL,
  keep_hdf5 = TRUE,
  acq2hdf5_location = NULL,
  ...
)
}
\arguments{
\item{acq_files}{character vector of acq files to be converted to HDF5, then imported as data.frame}

\item{hdf5_output_dir}{location for converted hdf5 files produced by acq2hdf5. Passed to acq2hdf5.}

\item{keep_hdf5}{logical indicating whether to retain hdf5 file after import completes. Default: TRUE}

\item{acq2hdf5_location}{path to acq2hdf5 binary, installed by the bioread python package.}
}
\value{
a list of
}
\description{
Import Acqknowledge ACQ files into data.frames
}
\details{
This function calls the acq2hdf5 wrapper R function. The converted HDF5 files are then
  parsed by biopac_hdf5_to_dataframe, which relies heavily on the data.table package to build the
  data.frame containing physiological data.
}
