% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_peaks.R
\name{import_wfdb_annotations}
\alias{import_wfdb_annotations}
\title{Helper function to read single-channel WFDB annotations into a data.table object}
\usage{
import_wfdb_annotations(
  wfdb_file,
  annotator,
  elapsed = TRUE,
  channel = 0L,
  type = "N",
  wfdb_path = "~/Downloads/wfdb/bin"
)
}
\arguments{
\item{wfdb_file}{Prefix to wfdb file containing the annotations to be extracted}

\item{annotator}{The algorithm used for creating the annotations
(determines the suffix, like 100.gqrs)}

\item{elapsed}{A boolean indicating whether to extract elapsed times since beginning of record
or absolute times. A value of TRUE passes -e to rdann}

\item{channel}{optional channel to read annotations from if file contains multi-channel data.
Default is 0, which is the first channel.}

\item{type}{The type of annotation to be imported from WFDB. Defaults to "N" for normal beats in ECG}

\item{wfdb_path}{location of the WFDB bin directory on this computer.}
}
\description{
Helper function to read single-channel WFDB annotations into a data.table object
}
\author{
Michael Hallquist
}
