% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-cluster.R
\name{init.clusters}
\alias{init.clusters}
\title{Initialize clusters}
\usage{
init.clusters()
}
\description{
Initialize global clusters for parallel processing.
Utilizes configuration defined by config.yaml and system configurations
to define number of cores to utilize.
All values are given to the System global config in the NeuroMap environment.
}
