% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_log.R
\name{init_eyelog}
\alias{init_eyelog}
\title{Initializes .elog file which will store eye-specific messages about how QA checks are going.}
\usage{
init_eyelog(log_dir = NULL)
}
\description{
Initially, my thought is to create .blog and .plog files for behavior and physio respectively and eventually write a separate function (or set of functions) that searches through logs and documents how certain checks went in a more succinct format (e.g. a .csv with subjects on rows and stages of QA on columns). 
 A report can then be generated to either verify folks that look fine or to draw attention to problematic data.
}
