% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eye.R
\name{initialize_eye}
\alias{initialize_eye}
\title{generic function for doing initial test on eye data.}
\usage{
initialize_eye(eye)
}
\description{
This includes validation of very basic data quality (large variance in gaze distribution, excessive blinks, large jumps in eye position, etc).
TODO: include functionality for logging of successes, warnings, failures. This will probably involve a trycatch statement that could handle a potentially large number of issues. We'll have to see how complicated it gets by balancing flexibility with parsimony. Tend to prefer flexibility if the package allows user-side functionality to be parsimonious :)
TODO: perhaps even store key variables (e.g. some measure of pupil fluctuation, or saccade velocity/acceleration) from prior subjects in separate circumscribed csv (which values get appended to) and plot distributions for every new subject. This would be akin to constructing a sort of empirical null distribution and performing informal (visual)"hypothesis tests" where we would hope certain variables in a given subject are not "significantly different" than the group distribution.
}
