% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-logger.R
\name{log.debug}
\alias{log.debug}
\title{Logs message as a debug using log object}
\usage{
\method{log}{debug}(msg, logger = "")
}
\arguments{
\item{msg}{Message to log}

\item{logger}{Logger object name}
}
\description{
Logs message as a debug using log object
}
