% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-logger.R
\name{log.info}
\alias{log.info}
\title{Logs message as an info using log object}
\usage{
\method{log}{info}(msg, logger = "")
}
\arguments{
\item{msg}{Message to log}

\item{logger}{Logger object name}
}
\description{
Logs message as an info using log object
}
