% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_log.R
\name{log_chunk_header}
\alias{log_chunk_header}
\title{Print a standardized chunk header}
\usage{
log_chunk_header(text)
}
\arguments{
\item{text}{Text to print into chunk header.}
}
\description{
Prints log chunks (to be run at the top of every sub-function of ep.eye_process_subject.R)
}
\note{
This simply standardizes for later crawler scripts to recognize.
}
