% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physio.R
\name{read_acq_txt}
\alias{read_acq_txt}
\title{Read a text file generated by acq2txt}
\usage{
read_acq_txt(acq_txt_file, ...)
}
\arguments{
\item{acq_txt_file}{A .txt or .txt.gz file created by acq2txt}
}
\description{
At present, this is just a thin wrapper around fread, which reads
  in the Acqknowledge txt file of the physio acquisition. Given that the acq2hdf5
  approach seems both faster and more robust, I've developed that in greater depth
  and this remains as more of a useful stub
}
\examples{
\dontrun{
  acq_data <- read_acq_txt("~/temp_acq/nmap016.txt.gz", na.strings=".")
}
}
