% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_behav.R
\name{read_behav}
\alias{read_behav}
\title{general wrapper for reading behavioral files into the package}
\usage{
read_behav(file, parser = NULL, ...)
}
\description{
general wrapper for reading behavioral files into the package
}
