% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_behav.R
\name{read_behav_vending}
\alias{read_behav_vending}
\title{Import Quail, Morris, Balleine Vending Machine task data}
\usage{
read_behav_vending(files = NULL)
}
\arguments{
\item{files}{A names character vector of three CSV files. Elements should be named 'pav', 'ins', 'pit'}
}
\description{
Import Quail, Morris, Balleine Vending Machine task data
}
\examples{
\dontrun{
  b <- read_behav_vending(c(
     pav="070_Pavlovian_2020_Feb_18_1421.csv",
     ins="070_Instrumental_2020_Feb_18_1412.csv",
     pit="070_transfer_2020_Feb_18_1438.csv")
  )
}
}
