% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_edf}
\alias{read_edf}
\title{Import EDF File}
\usage{
read_edf(
  edf_files,
  asc_output_dir = NULL,
  keep_asc = TRUE,
  gzip_asc = TRUE,
  samples = TRUE,
  ...
)
}
\arguments{
\item{edf.file}{Path to *edf file in UTF-8 format (UNIX/MAC pathing).  Paths will
be converted relative to your OS.}
}
\value{
list of subject eye data, where each element is a list generated by eyelinker::read.asc
}
\description{
A convience wrapper around FDBeye file importer to
     import using UTF-8 format
}
\details{
Call edf2asc" R function to conver *edf files to *edf
    files.  This wrapper  only allows a single path instead of multiple.
    Existing *asc files will be overwritten because we se the option to -y.
}
\examples{
\dontrun{
  eye_parsed <- read_edf(c("~/Box/s3_behav_data/neighborhood/eye/data/raw/N_003_TS.edf",
      "~/Box/s3_behav_data/neighborhood/eye/data/raw/N_004_az.edf"), keep_asc=FALSE, parse_all=TRUE)
}
}
