% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eye.R
\name{read_eye_neighborhood}
\alias{read_eye_neighborhood}
\title{specific function for importing neighborhood eye tracking data into the package
need to add task-specific processing}
\usage{
read_eye_neighborhood(file)
}
\description{
specific function for importing neighborhood eye tracking data into the package
need to add task-specific processing
}
