% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_physio}
\alias{read_physio}
\title{general wrapper for reading physio data into the package}
\usage{
read_physio(file, parser = NULL, ...)
}
\description{
general wrapper for reading physio data into the package
}
