% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physio.R
\name{read_physio_neighborhood}
\alias{read_physio_neighborhood}
\title{Read physiology from neighborhood task}
\usage{
read_physio_neighborhood(physio_data)
}
\arguments{
\item{physio_data}{This can be an \code{ep.physio} object, in which case we just window it to the relevant
time periods. Or, this}
}
\description{
Read physiology from neighborhood task
}
