% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eye.R
\name{read_process_eye}
\alias{read_process_eye}
\title{general wrapper for reading eye data into the package and performing all QA and processing}
\usage{
read_process_eye(file, yaml = NULL, gen_log = TRUE, log_dir = NULL, ...)
}
\description{
general wrapper for reading eye data into the package and performing all QA and processing
}
