% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{run.block}
\alias{run.block}
\title{Runs a block and follows the output until completion}
\usage{
run.block(block, input = list())
}
\arguments{
\item{block}{Block to run}

\item{input}{Input of block (this should only be used during recursion)}
}
\value{
0 if succesful, 1 if error
}
\description{
Recursive function to follow blocks until no more outputs.
}
