% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{set.block.name}
\alias{set.block.name}
\title{Set name of block object}
\usage{
set.block.name(block, block.name)
}
\arguments{
\item{block}{Block object}

\item{block.name}{name of block}
}
\description{
Set name of block object
}
