% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{set.callback}
\alias{set.callback}
\title{Set callback function of block object}
\usage{
set.callback(block, callback)
}
\arguments{
\item{block}{Block object}

\item{callback}{callback function of block}
}
\description{
Set callback function of block object
}
