% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{set.input.block}
\alias{set.input.block}
\title{Set input of block object}
\usage{
set.input.block(block, input.blocks)
}
\arguments{
\item{block}{Block object}

\item{input.blocks}{inputs of block}
}
\description{
Set input of block object
}
