% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap-block.R
\name{set.output.blocks}
\alias{set.output.blocks}
\title{Set output of block object}
\usage{
set.output.blocks(block, output.blocks)
}
\arguments{
\item{block}{Block object}

\item{output.blocks}{outputs of block}
}
\description{
Set output of block object
}
