% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-cluster.R
\name{terminate.clusters}
\alias{terminate.clusters}
\title{Terminates clusters}
\usage{
terminate.clusters()
}
\description{
Closes sockit to http port to stop all parallel processing using
the NeuroMap cluster.
Note that this function exists to explicitly terminate the cluster.
}
