% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_log.R
\name{tryCatch.ep}
\alias{tryCatch.ep}
\title{TryCatch function to run subroutines with ease and standardized output.}
\usage{
tryCatch.ep(code, describe_text = NULL, ...)
}
\arguments{
\item{code}{Chunk of code for tryCatch to evaluate. As in the TC documentation, multiple lines of code should be contained within curly brackets[{}].}

\item{describe_text}{String containing standardized information about the significance of the code being run. This will print as COMPLETE for successful execution and ERROR and WARNING is something undesirable happens, while allowing code execution to continue below.}
}
\description{
This tryCatch wrapper prints errors and warnings alongside information describing the process being executed. If code runs succesfully will print COMPLETE. These errors/warnings are ideally passed to a log file that documents sequential steps in a pipeline.
}
