% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_exp_yaml.R
\name{validate_exp_yaml}
\alias{validate_exp_yaml}
\title{Validate experimental config file}
\usage{
validate_exp_yaml(yaml_file)
}
\arguments{
\item{yaml_file}{path to YAML config file (see vignette on how to set up your ep.eye config file).}
}
\description{
Validate experimental config file
}
\author{
Michael Hallquist
}
