% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_preprocess_helpers.R
\name{ep.eye_collapse_time}
\alias{ep.eye_collapse_time}
\title{collapse timing to one row per time (paste multiple messages in a single et.msg)}
\usage{
ep.eye_collapse_time(ep.eye, dt)
}
\description{
collapse timing to one row per time (paste multiple messages in a single et.msg)
}
