% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_inherit_btw_ev}
\alias{ep.eye_inherit_btw_ev}
\title{Interface with messages passed with not recording}
\usage{
ep.eye_inherit_btw_ev(ep.eye, inherit_btw_ev, dt = NULL)
}
\arguments{
\item{ep.eye}{An initialized ep.eye object}

\item{inherit_btw_ev}{A named list taken from your config file. \code{inherit_btw_ev$calibration_check} can contain \code{$cal} and \code{$val} which are character vectors to search for in event .5 (prior to any recording event) and are checked to make sure GOOD is appended to these messages to ensure calibration and validation did not encounter any errors. The second element \code{inherit_btw_ev$calibration_check[[str/align_msg/pre_post]]} provides the opportunity to pass strings (\code{$str}) to search for amongst between-event messages, messages within events to align these messages to (\code{$align_msg}), and whether to assign between-event messages to the "pre" (e.g. moving message in 1.5 to 1) or "post" (e.g. moving message in 1.5 to 2) event (\code{$pre_post}). Elements of \code{$move_to_within} must be of the same length and will enounter error if this is not the case.}

\item{dt}{descriptive text to print to log file, defaults to NULL.}
}
\value{
ep.eye
}
\description{
"Between-event messages" are messages passed to the eyetracker while the system is not actively recording. Sometimes these contain event-relevant information that you would like to pull into a recording event itself (for example, a "trial ID" message that gets passed right before the recording starts). This function allows for 1) checking of calibration and validation messages that are passed before any actual recording events are apparent in the data ("calibration_check") and 2) specific messages to be integrated into the event data depending on "move_to_within
}
\author{
Nate Hall
}
