% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize.R
\name{ep.eye_initialize}
\alias{ep.eye_initialize}
\title{generic function for initializing ep.eye object and performing basic internal checks on the eye data, while remaining agnostic to task/behavior structure.}
\usage{
ep.eye_initialize(
  file,
  expected_edf_fields = c("raw", "sacc", "fix", "blinks", "msg", "input", "button",
    "info", "asc_file", "edf_file"),
  task = NULL,
  gaze_events = c("sacc", "fix", "blink"),
  meta_check = NULL,
  inherit_btw_ev = NULL,
  header = "2. Initialize eye object:",
  ...
)
}
\arguments{
\item{eye}{raw eye object pulled directly from the .edf file using read_edf(). Must be a list with expected_edf_fields c("raw", "sacc", "fix", "blinks", "msg", "input", "button", "info", "asc_file", "edf_file").}
}
\description{
This includes validation of very basic data quality (large variance in gaze distribution, excessive blinks, large jumps in eye position, etc).
TODO: include functionality for logging of successes, warnings, failures. This will probably involve a trycatch statement that could handle a potentially large number of issues. We'll have to see how complicated it gets by balancing flexibility with parsimony. Tend to prefer flexibility if the package allows user-side functionality to be parsimonious :)
TODO: perhaps even store key variables (e.g. some measure of pupil fluctuation, or saccade velocity/acceleration) from prior subjects in separate circumscribed csv (which values get appended to) and plot distributions for every new subject. This would be akin to constructing a sort of empirical null distribution and performing informal (visual)"hypothesis tests" where we would hope certain variables in a given subject are not "significantly different" than the group distribution.
}
