% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events_helpers.R
\name{ep.eye_parse_event_info}
\alias{ep.eye_parse_event_info}
\title{split off function for extracting event info}
\usage{
ep.eye_parse_event_info(
  ep.eye,
  extraction_method,
  extract_event_func_path,
  csv_path,
  msg_seq,
  dt
)
}
\description{
split off function for extracting event info
}
