% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_parse_events.R
\name{ep.eye_parse_events}
\alias{ep.eye_parse_events}
\title{Parse task config file, extracting essential task information}
\usage{
ep.eye_parse_events(
  ep.eye,
  extract_event_func_path,
  csv_path,
  msg_seq,
  header = NULL
)
}
\arguments{
\item{ep.eye}{ep.eye object that has been previously initialized and tidied}

\item{extract_event_func_path}{Path to user-defined message parsing function here}

\item{csv_path}{Path to write event .csvs to}

\item{msg_seq}{List of optional message sequence arguments passed in config file. Can contain \code{msg_start}, \code{msg_end}, \code{eval_middle}, \code{ordered}. See the ep.eye_config vignette for details on these fields.}

\item{header}{String for section header. Defaults to NULL}
}
\value{
ep.eye appended with important event-level information included.
}
\description{
Parse task config file, extracting essential task information
}
\author{
Nate Hall
}
