% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_rm_crinfo}
\alias{ep.eye_rm_crinfo}
\title{Remove useless cr.info column.}
\usage{
ep.eye_rm_crinfo(ep.eye)
}
\arguments{
\item{ep.eye}{An ep.eye object.}
}
\value{
ep.eye
}
\description{
Removes cr.info column if it contains unhelpful information
}
