% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.eye_initialize_helpers.R
\name{ep.eye_shift_timing}
\alias{ep.eye_shift_timing}
\title{Shift timing of ep.eye files to 0 start point}
\usage{
ep.eye_shift_timing(ep.eye, dt)
}
\description{
Shift timing of ep.eye files to 0 start point
}
