% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_edf}
\alias{read_edf}
\title{Import EDF File}
\usage{
read_edf(
  edf_files,
  asc_output_dir = NULL,
  keep_asc = TRUE,
  gzip_asc = TRUE,
  samples = TRUE,
  header = NULL,
  parse_all = TRUE,
  ...
)
}
\arguments{
\item{samples}{Logical. Whether to import raw data at the sampling rate aquired off the eyetracker.}

\item{parse_all}{Logical. Whether to additionally parse between-event messages.}

\item{edf.file}{Path to *edf file in UTF-8 format (UNIX/MAC pathing).  Paths will be converted relative to your OS.}
}
\value{
list of subject eye data, where each element is a list generated by eyelinker::read.asc
}
\description{
A convience wrapper around FDBeye file importer to
     import using UTF-8 format
}
\details{
Call edf2asc" R function to convert *edf files to *asc
    files.  This wrapper  only allows a single path instead of multiple.
    Existing *asc files will be overwritten because we se the option to -y.
    Before using this function, you must have the SR Research edf2asc downloaded from the SR Research Website (Downloads/Eyelink Developers Kit: https://www.sr-support.com/thread-13.html)
    and included in PATH. See https://rdrr.io/github/davebraze/FDBeye/man/edf2asc.html for details on how to add the command-line utility to your PATH.
}
\examples{
\dontrun{
  eye_parsed <- read_edf(c("~/Box/s3_behav_data/neighborhood/eye/data/raw/N_003_TS.edf",
      "~/Box/s3_behav_data/neighborhood/eye/data/raw/N_004_az.edf"), keep_asc=FALSE, parse_all=TRUE)
}
}
